% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{addAllGenesNotInTiMEx}
\alias{addAllGenesNotInTiMEx}
\title{Appends the genes as single clusters}
\usage{
addAllGenesNotInTiMEx(resIterTiMEx)
}
\arguments{
\item{resIterTiMEx}{structure containing the results from iteratively
running TiMEx on a binary dataset, as returned by the function
\code{iterTiMExSims}.}
}
\value{
groupysL list with the identified clusters. each element is a cluster
with gene indices.
}
\description{
\code{addAllGenesNotInTiMEx} appends the genes not indentified
in any mutually exclusive groups as clusters of single genes.
}
\details{
The function returns the updated clustering which also includes
the single genes.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

