% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsTiMEx.R
\name{addNoise}
\alias{addNoise}
\title{Add noise to a binary matrix}
\usage{
addNoise(mat, epsilon)
}
\arguments{
\item{mat}{binary input matrix, with rows representing samples and columns
representing genes}

\item{epsilon}{the error rate at the level of genes, real number 
between 0 and 1}
}
\value{
The input binary matrix with added noise.
}
\description{
\code{addNoise} adds binomially sampled noise to a binary 
matrix, with success probability epsilon.
}
\details{
Any binary matrix can be used as input.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

