% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsOth.R
\name{addPathBinary}
\alias{addPathBinary}
\title{Summarizes the inferred groups}
\usage{
addPathBinary(resIterBinTiMEx)
}
\arguments{
\item{resIterBinTiMEx}{structure with the results of running TiMEx 
iteratively on any input binary matrix, as returned by the function
\code{iterTiMExBinaryMats}.}
}
\value{
resIterBinTiMEx, the input structure, with the added field 
\code{genesBinary}. \code{genesBinary} is a matrix with rows the 
identified mutually exclusive pathways and columns the genes. Each entry 
[i,j] = 1 if gene j was identified in pathway i and 0 otherwise.
}
\description{
\code{addPathBinary} summarizes the assignment of genes to 
pathways as a binary matrix.
}
\details{
This function appends to the input structure the field 
\code{genesBinary}, representing the assignment of the identified genes to 
pathway, in the form of a binary matrix.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

