% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSearch.R
\name{cleanGroupys}
\alias{cleanGroupys}
\title{Removes empty groups}
\usage{
cleanGroupys(groupysInput)
}
\arguments{
\item{groupysInput}{list with as many elements as mutually exclusive groups,
each element consists of the genes assigned to the respective pathway. It
possibily contains empty elements.}
}
\value{
list with as many elements as mutually exclusive groups,
each element consists of the genes assigned to the respective pathway. It
doesn't contain any empty elements.
}
\description{
\code{cleanGroupys} removes the empty elements in a list, 
representing the assignment of genes to pathways.
}
\details{
The empty elements are a consequence of the optimization step, in
which a grouping with less pathways than initially attains the minimal 
score.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

