% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{clustMetrIterTiMEx}
\alias{clustMetrIterTiMEx}
\title{Assesses the performance of the inferred groups after running 
TiMEx iteratively}
\usage{
clustMetrIterTiMEx(resIterTiMEx)
}
\arguments{
\item{resIterTiMEx}{structure containing the results from iteratively
running TiMEx on a binary dataset, as returned by the function
\code{iterTiMExSims}.}
}
\value{
list consisting of 
\itemize{
\item{\code{reconstruction}} {structure with metrics of the reconstuction of
the initial groups, as follows
\itemize{
\item{\code{groupsSpread}}{list containing as many elements as real 
pathways. each element is a list with at most as many elements as the 
number of inferred pathways (including potential NULL elements). 
each element of this second list  is a vector of genes part of the current
real pathway and also part of the inferred pathway. In othder words, the 
first index represents the real group and the second index represents the 
inferred group.}
\item{\code{groupsSpreadPercent}} {list containing as many elements as real 
pathways. each element is a list with at most as many elements as the 
number of inferred pathways (including potential NULL elements). 
each element of this second list  is the percentage of 
reconstruction of the respective first pathway by the genes identified as 
part of the second pathway. In othder words, the first index represents 
the real group and the second index represents the inferred group.}
\item{\code{groupsSpreadOrd}} {list containing as many elements as real 
pathways, ordered by the real mu values of the pathways. each element is a 
list with at most as many elements as the number of inferred pathways 
(including potential NULL elements). each element of this second list is a 
vector of genes part of that particular real pathway 
and also inferred as part of the current pathway. In othder words, the 
first index represents the real group and the second index represents the 
inferred group.}
\item{\code{groupsSpreadPercentOrd}} {list containing as many elements as 
real pathways, orderde by the real mu values of the pathways. 
each element is a list with at most as many elements as the 
number of inferred pathways (including potential NULL elements). each 
element of this second list  is the percentage of reconstruction of the 
respective first pathway by the genes identified as part of the second 
pathway. In othder words, the first index represents the real group and the 
second index represents the inferred group.}
\item{\code{groupsCompact}} {list containing as many elements as real 
pathways. each element is a vector consisting of the genes members of
that particular pathway which were identified as part of a mutually 
exclusive group, regardless of the group in which they were identified.}
\item{\code{groupsCompactPercent}} {list containing as many elements as real 
pathways. each element is the percentage of reconstruction of that 
particular pathway, i.e. the percentage of genes which were identified as 
mutually exclusive, regardless of the group in which they were identified.}
\item{\code{groupsCompactOrd}} {list containing as many elements as real 
pathways, ordered by the real mu values of the pathways. each element is a 
vector consisting of the genes members of that particular pathway which 
were identified as part of a mutually exclusive group, regardless of the 
group in which they were identified.}
\item{\code{groupsCompactPercentOrd}} {list containing as many elements as 
real pathways, ordered by the real mu values of the pathwayss. each 
element is the percentage of reconstruction of that particular pathway, 
i.e. the percentage of genes which were identified as mutually exclusive, 
regardless of the group in which they were identified.}
\item{\code{groupsEmpty}} {list containing as many elements as 
real pathways. each element contains the genes from that particular pathway
which were not identified as part of mutually exclusive groups.}
\item{\code{groupsEmptyPercent}} {list containing as many elements as 
real pathways. each element is the percentage of the respective pathway 
which was not identified as part of mutually exclusive groups.}
\item{\code{groupsEmptyOrd}} {list containing as many elements as 
real pathways, ordered by the real mu values of the pathways. each element 
contains the genes from that particular pathway which were not identified 
as mutually exclusive.}
\item{\code{groupsEmptyPercentOrd}} {list containing as many elements as 
real pathways, ordered by the real mu values of the pathways. each element 
is the percentage of the respective pathway which was not identified as 
part of mutually exclusive groups.}
\item{\code{booleanOthers}} {list containing as many elements as 
real pathways. each element is a list with at most as many elements as the 
number of inferred pathways (including potential NULL elements). 
each element of this second list is either 0, if the 
inferred group contained elements from more than one real pathway, or 1, if
the inferred group only contained elements from one real pathway. In othder 
words, the first index represents the real group and the second index 
represents the inferred group.}
\item{\code{booleanOthersOrd}}{list containing as many elements as 
real pathways, orderde by the real mu values of the pathways. each element 
is a list with at most as many elements as the number of inferred pathways 
(including potential NULL elements). each element of this second list is 
either 0, if the inferred group contained elements from more than one real 
pathway, or 1, if the inferred group only contained elements from one real 
pathway. In othder words, the first index represents the real group and the 
second index represents the inferred group.}
}}
\item{\code{characteristics}} {structure with metrics per pathway 
characterizing the identified groups, as follows
\itemize{
\item{\code{freqGenes}} {list containing as many elements as real pathways.
each element is a vector with frequencies of the genes inside that pathway.}
\item{\code{freqGenesAvg}} {list containing as many elements as real 
pathways. each element is the average frequency of all genes inside that 
pathway (averaged over all the genes in each pathway).}
\item{\code{freqGenesOrd}} {list containing as many elements as real 
pathways. each element is a vector with the frequencies of the 
genes inside that pathway, ordered decreasingly by the intensities of 
mutual exclusivity of each pathway.}
\item{\code{realMus}} {vector containing the intensities of mutual 
exclusivity of the real pathways}
\item{\code{realMusOrd}} {vector containing the intensities of mutual 
exclusivity of the real pathways, ordered decreasingly}
}}
\item{\code{peredge}} {structure with metrics per edge characterizing the 
reconstruction of groups, as follows
\itemize{
\item{\code{tp}} {true positive rate: the number of true mutual exclusivity
pairwise connections which were also identified, divided by the total 
number of true mutual exclusivity pairwise connections.}
\item{\code{fp}} {false positive rate: the number of false mutual 
exclusivity edges identified, divided by the total number of possible
pairwise connections.}
}}
\item{\code{resIterTiMEx}} {input structure containing the identified 
mutually exclusive groups after running TiMEx itartively, together with 
the simulated data on the basis of which the inference was done.}
}
}
\description{
\code{clustMetrIterTiMEx} computes per-edge and per-pathway 
metrics evaluating the performance of the pathway assignment inferred after
running TiMEx iteratively, based on data simulated from PathTiMEx
}
\details{
This function assesses how well the real groups were recovered 
after performing iterative inference with TiMEx.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

