% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsCBN.R
\name{compareStructs}
\alias{compareStructs}
\title{Compares two posets}
\usage{
compareStructs(oldStruct, newStruct)
}
\arguments{
\item{oldStruct}{the first of the two structures to compare, corresponding to
the previous optimization step.}

\item{newStruct}{the second of the two structures to compare, corresponding
to the current optimization step.}
}
\value{
list consisting of 
\itemize{
\item{\code{falseNeg}} {number of edges which are in the old structure,
but not in the new structure, divided by the number of all possible directed
edges.}
\item{\code{falsePos}} {number of edges which are in the new structure,
but not in the old structure, divided by the number of all possible directed
edges.}
\item{\code{accuracy}} {measure of similarity of the two structures, equal
to 1 - number of mismatches divided by the number of all possible directed 
edges. In case the two structures are equal, \code{accuracy = 1}.}
\item{\code{noMismatches}} {the number of edges which are found in only one
of either of the two structures}
\item{\code{densOld}} {number of edges in the old structure.}
\item{\code{densNew}} {number of edges in the new structure.}
}
}
\description{
\code{compareStructs} compares two posets of same size, 
encoded as binary matrices. It returns various metrics comparing the two
matrices.
}
\details{
For computing the false posive and false negative rates, it is 
assumed that the first argument (oldStruct) is the real structure. In this
help page, fhe first structure is referred to as the old structure, while 
the new structure is referred to as the new structure. The number of all 
possible directed edges, often used to compute rates, equals twice the 
number of events choose 2. In case the two structures are equal, the 
\code{accuracy} measure is 1.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

