% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSearch.R
\name{computeScore}
\alias{computeScore}
\title{Computes clustering score}
\usage{
computeScore(Datamat, groupys, adjmat)
}
\arguments{
\item{Datamat}{binary alteration matrix where rows are samples and columns
are genes and an 1 on position [i,j] means that gene j is altered in sample
i.}

\item{groupys}{list with as many elements as mutually exclusive groups,
each element consists of the genes assigned to the respective pathway.}

\item{adjmat}{binary matrix, representing the optimal structure which is
kept fixed. The matrix contains one dummy column and row at the end, 
identically 0, hence its dimension is number of groups +1.}
}
\value{
clustering score (integer)
}
\description{
\code{computeScore} computes the score of a given clustering, 
for a given structure.
}
\details{
The score equals the number of contradictions due to both mutual 
exclusivity and progression, or the number of ones that would need to be 
changed to zeroes to ensure consistency of the data with the clusters and 
structure.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

