% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsOth.R
\name{count.duplicates}
\alias{count.duplicates}
\title{Counts duplicates in a data frame}
\usage{
count.duplicates(DF)
}
\arguments{
\item{DF}{the data frame whose rows will be counted}
}
\value{
a data frame with unique rows, followed by an additional column
which counts the number of times each row was found in the input data
frame.
}
\description{
\code{count.duplicates} counts how many times each row appears
in the input data frame
}
\details{
This function returns a new data frame, with the unique rows of the
input one, and an additional column, counting how many times each row 
appears.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

