% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsOth.R
\name{countPerfSims}
\alias{countPerfSims}
\title{Assesses how well each group was reconstructed}
\usage{
countPerfSims(pathwaysTable, groups)
}
\arguments{
\item{pathwaysTable}{a binary matrix of all groups identified aross many
repetitions, as returned by multiple runs of \code{addPathBinary}. rows 
represent indetified groups and columns represent genes. the element [i,j] 
is 1 if gene j was identified as part of pathway i and 0 otherwise.}

\item{groups}{a list with the real assignment of genes to pathways. each 
element of the list is a vector with the genes belonging to that pathway.}
}
\value{
a binary matrix in which the first columns represent the unique 
pathways identified in the input matrix (columns named with \code{Gr}, 
followed by the group index, followed by \code{ge}, followed by the gene
index). next column, \code{counts}, counts how many times each group was 
identified as mutually exclusive. next columns correspond to each of the 
unique pathways identified, and show either the total number of genes form 
each group which were part of the group (\code{Tot}), or how much this 
number represents, as a percentage, from the real number of genes in each 
group (\code{Per}).
}
\description{
\code{countPerfSims} assesses the reconstruction performance
of each group, inferred from any binary matrix.
}
\details{
This function summarizes multiple repetitions of the inference
of pathways from a simulated dataset. Ideally, the groups with largest 
counts are the real groups.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

