% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsCBN.R
\name{doPatForCBN}
\alias{doPatForCBN}
\title{Produces \code{pat} files for H-CBN}
\usage{
doPatForCBN(groupsList, genesMat, write, path, name)
}
\arguments{
\item{groupsList}{list with as many elements as mutually exclusive groups,
each element consists of the genes assigned to the respective pathway}

\item{genesMat}{binary alteration matrix with rows representing patients 
and columns representing genes, where an 1 indicates that a gene has been
altered in one patient}

\item{write}{boolean, indicating whether to also write the output matrix to 
a file. Default to FALSE.}

\item{path}{path where the output matrix should be written. Default to
"~/Dropbox/PathTiMEx/RPathTiMEx/txts".}

\item{name}{name of the file to be written. Default to "outCBN".}
}
\value{
binary alteration matrix representing the alteration of pathways
in patients.
}
\description{
\code{doPatForCBN} produces a binary matrix with dimensions 
patients x pathways, to be used as input by H-CBN. It can also write the 
matrix to a file.
}
\details{
The matrix returned by this function is the binary alteration matrix
on the basis of which order constraints between pathways are inferred. The
input to this function are the assignment of genes to pathways (as a list)
and the binary alteration matrix on the level of genes. A pathway is 
considered to be altered as soon as at least one of its gene members is 
altered. If \code{write=TRUE} and if \code{path} and \code{name} are 
provided, the matrix is written to the file \code{path/name.pat}.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

