% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{drawSamples}
\alias{drawSamples}
\title{Draws samples from a pathTiMEx model}
\usage{
drawSamples(m, model)
}
\arguments{
\item{m}{number of samples to be simulated}

\item{model}{the pathTiMEx model, as a list, as returned by the functions 
which generate data: \code{generateRandomModel}, \code{generateFixedModel} 
and \code{generateStructureModel}}
}
\value{
list consisting of
\itemize{
\item{\code{XGenes}} {binary matrix with the simulated observations on the
level of genes, including noise. Rows represent samples and columns 
represent genes. }
\item{\code{XGenesClean}} {binary matrix with the simulated observations on
the level of genes, before noise was added. Rows represent samples and 
columns represent genes.}
\item{\code{XPathwaysClean}} {binary matrix with the simulated observations
on the level of pathways, before noise was added. Rows represent samples  
and columns represent pathways.}
\item{\code{TGenes}} {matrix with the waiting times of the genes. Rows 
represent samples and columns represent genes.}
\item{\code{TPathways}} {matrix with the waiting times of the pathways. The 
waiting time of each pathway equals the minimum waiting time of its gene 
members. Rows represent samples and columns represent pathways.}
\item{\code{Tobs}} {vector with the observation times. For each sample,
an independent observation time is drawn exponentially.}
\item{\code{ZGenes}} {matrix with the independent exponential components 
contributing to the waiting times of the genes. Rows represent samples 
and columns represent genes.}
\item{\code{m}} {number of samples}
\item{\code{model}} {structure with details about the model, s.a. waiting 
time rates, poset etc. This type of structure is returned by the functions
whcih generate pathTiMEx models: \code{generateRandomModel}, 
\code{generateFixedModel} and \code{generateStructureModel}.}
}
}
\description{
\code{drawSamples} generates samples form a pathTiMEx model
}
\details{
This function generates a binary matrix on the level of genes 
from the pathTiMEx model, as well as various other features of the model, 
such as the waiting times to alteration, the observation times etc.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

