% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{generateFixedModel}
\alias{generateFixedModel}
\title{Generates a pathTiMEx model for given structure and parameters}
\usage{
generateFixedModel(lams, lamobs, mus, listEdges, assignment, epsilon)
}
\arguments{
\item{lams}{the exponential rates of the waiting times to mutations}

\item{lamobs}{the exponential rate of the observation time (usually fixed
to 1)}

\item{mus}{the intensities of mutual exclusivity}

\item{listEdges}{list of pairs [i,j] such that the element [i,j] in the 
poset equals 1}

\item{assignment}{vector of assignments of genes to pathways}

\item{epsilon}{{the error rate at the level of genes, real number 
between 0 and 1}}
}
\value{
lists of parameters and structure for a pathTiMEx model, as follows:
\itemize{
\item{\code{lams}} {input vector of exponential rates of the waiting times 
to mutation of each gene}
\item{\code{mus}} {input vector of mutual exclusivity intensities of each 
pathway}
\item{\code{poset}} {the poset of the model}
\item{\code{assignment}} {the assignment of genes to pathways}
\item{\code{epsilon}} {the error rate at the level of genes, real number 
between 0 and 1}
\item{\code{type}} {way in which the model was simulated, in this case 
\code{fixed}}
}
}
\description{
\code{generateFixedModel} generates a pathTiMEx model 
for given structure (poset and pathway assignment) and parameters
}
\details{
The exponential rate of the observation time is kept fixed (usually
to 1) when generating parameters for the model
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

