% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{generateRandomModel}
\alias{generateRandomModel}
\title{Randomly generates a pathTiMEx model}
\usage{
generateRandomModel(n, k, lamobs, minLam, maxLam, muInterval, ldensesInterval,
  epsilon)
}
\arguments{
\item{n}{number of genes}

\item{k}{number of pathways}

\item{lamobs}{exponential rate of the observation time (usually fixed
to 1)}

\item{minLam}{lower limit for simulating the exponential rates of the
waiting times to mutations of the genes}

\item{maxLam}{upper limit for simulating the exponential rates of the
waiting times to mutations of the genes}

\item{muInterval}{vector of values from which the intensities of mutual
exclusivity of the pathways are uniformly sampled}

\item{ldensesInterval}{vector of values from which the density of edges in 
the poset is uniformly sampled}

\item{epsilon}{the error rate at the level of genes, real number between 0 
and 1}
}
\value{
lists of parameters and structure for a pathTiMEx model, as follows:
\itemize{
\item{\code{lams}} {vector of exponential rates (lambda) of the waiting
times to mutations of the genes}
\item{\code{lamobs}} {the exponential rate of the observation time, usually
fixed to 1}
\item{\code{mus}} {vector of intensities of mutual exclusivity (mu) for 
each pathway}
\item{\code{ldense}} {the density of edges in the poset}
\item{\code{poset}} {the poset of the model}
\item{\code{assignment}} {the assignment of genes to pathways}
\item{\code{epsilon}} {the error rate at the level of genes, real number 
between 0 and 1}
\item{\code{type}} {way in which the model was simulated, in this case 
\code{random}}
}
}
\description{
\code{generateRandomModel} generates randomly the structure 
and the parameters of a pathTiMEx model, including the exponential rates 
of the waiting times to mutations, the intensities of mutual exclusivity
of the pathways, the error rate, the poset and the pathway assignments.
}
\details{
The exponential rate of the observation time is kept fixed (usually
to 1) when generating parameters for the model.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

