% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{generateStructureModel}
\alias{generateStructureModel}
\title{Generates a pathTiMEx model for given parameters}
\usage{
generateStructureModel(lams, lamobs, mus, epsilon)
}
\arguments{
\item{lams}{vector of exponential rates of the waiting times to mutations, 
equal in size to the number of genes}

\item{lamobs}{the exponential rate of the observation time (usually fixed
to 1)}

\item{mus}{vector of intensities of mutual exclusivity, equal in size to
the number of pathways}

\item{epsilon}{the error rate at the level of genes, real number 
between 0 and 1}

\item{ldensesInterval}{vector of values from which the density of edges in 
the poset is uniformly sampled}
}
\value{
lists of parameters and structure for a pathTiMEx model, as follows:
\itemize{
\item{\code{lams}} {input vector of exponential rates of the waiting times 
to mutation of each gene}
\item{\code{lamobs}} {input exponential rate of the observation time}
\item{\code{mus}} {input vector of mutual exclusivity intensities of each 
pathway}
\item{\code{ldense}} {the density of edges in the poset}
\item{\code{poset}} {the poset of the model}
\item{\code{assignment}} {the assignment of genes to pathways}
\item{\code{epsilon}} {the error rate at the level of genes, real number 
between 0 and 1}
\item{\code{type}} {way in which the model was simulated, in this case 
\code{structure}}
}
}
\description{
\code{generateStructureModel} randomly generates the
structure for a pathTiMEx model under given parameters, i.e. the 
poset and the pathway assignments are generated, while the 
exponential rates of the waiting times to mutations 
and of the observation time, the intensities of mutual exclusivity and the 
error rate are known.
}
\details{
The exponential rate of the observation time is kept fixed (usually
to 1) when generating parameters for the model.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

