% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsPoset.R
\name{genRandomPoset}
\alias{genRandomPoset}
\title{Generates a poset}
\usage{
genRandomPoset(k, ldenses, trans_reduced = TRUE)
}
\arguments{
\item{k}{number of events (pathways)}

\item{trans_reduced}{boolean indicating whether the poset should be a
transitive reduction (default to \code{TRUE})}

\item{ldense}{ratio between 0 and 1 indicating how dense (in terms of 
edges) the poset is}
}
\value{
a poset (a binary upper triangular matrix)
}
\description{
\code{genRandomPoset} generates a random poset (partial
order at the level of pathways) with a given number of events (pathways) 
and a given edge density.
}
\details{
The poset is a binary upper triangular matrix in which if 
element [i,j]=1, then there is an order relation between i and j, i.e. event
i happens before or at the same time as event j. The 
parameter \code{ldenses} represents the fraction (between 0 and 1) of 
edges from the maximal number of edges (which contains nr_muts choose 2 
edges) to be included in the poset.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

