% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsTiMEx.R
\name{iterTiMExBinaryMats}
\alias{iterTiMExBinaryMats}
\title{Interative TiMEx on any binary dataset}
\usage{
iterTiMExBinaryMats(mat, pairMu, pairPvalue, groupPvalue, corr, noRuns)
}
\arguments{
\item{mat}{binary input matrix, with rows representing samples and columns
representing genes}

\item{pairMu}{threshold on the pair-level mu (input to the function 
\code{TiMEx}). Default to 0.5.}

\item{pairPvalue}{threshold on the pair-level pvalue (input to the 
function \code{TiMEx}). Default to 0.01.}

\item{groupPvalue}{threshold on the group-level pvalue (input to the 
function \code{TiMEx}). Default to 0.1.}

\item{corr}{type of correction employed by TiMEx. It has to either be "bonf"
or "fdr". Default to "bonf".}

\item{noRuns}{maximum number of times to run the iterative procedure. 
Default to 10.}
}
\value{
list consisting of 
\itemize{
\item{\code{geneNames}} {list consisting of the names of the genes part
of the identified mutually exclusive groups. each element is a vector of 
gene names corresponding to one group.}
\item{\code{geneIndices}} {list consisting of the indices in the input
binary matrix of the genes part of the identified mutually exclusive 
groups. each element is a vector of gene indices corresponding to one 
group.}
\item{\code{mat}} {input binary matrix}
\item{\code{samplesModel}} {input structure containing the simulated 
pathTiMEx model, including the binary matrix used as input to the function
\code{TiMEx}.}
\item{\code{muEst}} {vector with the estimated mu for the identified 
mutually exclusive groups, of length the number of groups identified.}
\item{\code{corr}} {the input method for multiple testing correction used
by TiMEx.}
}
}
\description{
\code{iterTiMExBinaryMats} performs TiMEx iteratively on any 
binary dataset, by each time removing the most significant group of largest 
size and runnning TiMEx again.
}
\details{
This function reports the most significant mutually exclusive 
groups of largest size inferred iteratively, on the basis of any binary input
matrix. Once a group is reported, it is removed from the dataset and TiMEx is 
ran again. The groups are reported decreasingly by size.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

