% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{iterTiMExSims}
\alias{iterTiMExSims}
\title{Performs TiMEx iteratively on a simulated dataset}
\usage{
iterTiMExSims(samplesModel, pairMu, pairPvalue, groupPvalue, corr, noRuns)
}
\arguments{
\item{samplesModel}{structure as returned by the function \code{drawSamples},
containing the simulated pathTiMEx model, including the simulated binary 
matrices, the waiting times etc}

\item{pairMu}{threshold on the pair-level mu (input to the function 
\code{TiMEx}). Default to 0.5.}

\item{pairPvalue}{threshold on the pair-level pvalue (input to the 
function \code{TiMEx}). Default to 0.01.}

\item{groupPvalue}{threshold on the group-level pvalue (input to the 
function \code{TiMEx}). Default to 0.1.}

\item{corr}{type of correction employed by TiMEx. It has to either be "bonf"
or "fdr". Default to "bonf".}

\item{noRuns}{maximum number of times to run the iterative procedure. 
Default to 10.}
}
\value{
list consisting of 
\itemize{
\item{\code{geneNames}} {list consisting of the names of the genes part
of the identified mutually exclusive groups. each element is a vector of 
gene names corresponding to one group.}
\item{\code{geneIndices}} {list consisting of the indices in the input
binary matrix of the genes part of the identified mutually exclusive 
groups. each element is a vector of gene indices corresponding to one 
group.}
\item{\code{geneAssignments}} {list consisting of the initial assignment to
pathways of the genes in the identified groups. Each element is a vector of 
pathway indices corresponding to an identified mutually exclusive group. 
The names of each vector represent the indices of the genes in the binary 
input matrix.}
\item{\code{samplesModel}} {input structure containing the simulated 
pathTiMEx model, including the binary matrix used as input to the function
\code{TiMEx}.}
\item{\code{muEst}} {vector with the estimated mu for the identified 
mutually exclusive groups, of length the number of groups identified.}
\item{\code{corr}} {the input method for multiple testing correction used
by TiMEx.}
}
}
\description{
\code{iterTiMExSims} performs TiMEx iteratively on a simulated 
dataset, by each time removing the highest ranking group of largest size and 
runnning TiMEx again.
}
\details{
The groups reported are always the most significant ones of largest
size. Once a group is reported, it is removed from the dataset and TiMEx is 
ran again. The groups are reported decreasingly by size. This function is
designed to be applied on data simulated by the functions in this package.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

