% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsPoset.R
\name{make_linear_poset}
\alias{make_linear_poset}
\title{Simulates a linear poset}
\usage{
make_linear_poset(p)
}
\arguments{
\item{p}{the number of events}
}
\value{
a binary matrix with relationships between events representing
order constraints as a chain (linear poset).
}
\description{
\code{make_linear_poset} simulates a linear poset, for a given
number of events.
}
\details{
A linear poset is used as starting solution for simulated annealing
for CBN structure optimization.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

