% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSearch.R
\name{optGroups}
\alias{optGroups}
\title{Locally optimizes pathway assignment}
\usage{
optGroups(Datamat, groupys, gamma, numsave, skipsteps, adjmat)
}
\arguments{
\item{Datamat}{binary alteration matrix where rows are samples and columns
are genes and an 1 on position [i,j] means that gene j is altered in sample
i.}

\item{groupys}{list with as many elements as mutually exclusive groups,
each element consists of the genes assigned to the respective pathway.}

\item{gamma}{parameter used in optimization to tune whether the MCMC 
sampler: a larger value encourages the chain to explore each local minimum 
more thoroughly, while a lower value allows the chain to escape each local 
area more easily and explore more of the global space. Usually set to 0.5.}

\item{numsave}{number of minimal scores attained by the chains, to bs used 
for plotting.}

\item{skipsteps}{number representing once in how many iterations to save
the minimal score for plotting}

\item{adjmat}{binary square matrix, representing the optimal structure which 
is kept fixed. The matrix contains one dummy column and row at the end, 
identically 0, hence its dimension is number of groups +1.}
}
\value{
list consisting of 
\itemize{
\item{\code{mingroupys}} {optimized clustering solution, which minimizes the
number of contradictions due to both mutual exclusivity and progression: 
list with as many elements as mutually exclusive groups, each element 
consists of the genes assigned to the respective pathway. }
\item{\code{mingroupysscore}} {minimal score, namely the number of 
contradictions due to both mutual exclusivity and progression.}
\item{\code{accepted}} {number of times the proposed move was accepted.}
\item{\code{allscores}} {the scores attained at each \code{skipsteps} 
iterations.}
}
}
\description{
\code{optGroups} optimizes the assignment of genes to pathways,
for given fixed structure.
}
\details{
The optimization is done by via an MCMC chain, by minimizing the 
number of contradictions due to both mutual exclusivity and progression, 
as the number of ones that would need to be changed to zeroes to ensure 
consistency of the data with the clusters and structure. The number of
iterations in each chain equals \code{numsave*skipsteps}. For each run, 
opening a new mutually exclusive group is allowed.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

