\name{pathTiMEx-package}
\alias{pathTiMEx-package}
\alias{pathTiMEx}
\docType{package}
\title{
\packageTitle{pathTiMEx}
}
\description{
\packageDescription{pathTiMEx}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pathTiMEx}
\packageIndices{pathTiMEx}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pathTiMEx}

Maintainer: \packageMaintainer{pathTiMEx}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
