% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsPoset.R
\name{plotPoset}
\alias{plotPoset}
\alias{plot_poset}
\title{Plots a poset}
\usage{
plotPoset(poset, size = 12)
}
\arguments{
\item{poset}{the poset to be plotted, given as a binary matrix}

\item{size}{the size of the names of events, default to 12}
}
\value{
the plot of the input poset
}
\description{
\code{plotPoset} plots a poset given as input, by representing 
the causal relations between events.
}
\details{
The column names of the poset should be the names of the events, 
and they will be used for plotting.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

