% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSearch.R
\name{returnAssignment}
\alias{returnAssignment}
\title{Returns assignment to groups as vector}
\usage{
returnAssignment(groupysInput)
}
\arguments{
\item{groupysInput}{list with as many elements as mutually exclusive groups,
each element consists of the genes assigned to the respective pathway}
}
\value{
a vector where each position represents a gene and each entry 
represents the group to which it is assigned
}
\description{
\code{returnAssignment} returns the assignment to pathways of 
a list of gene groups, as a vector of groups.
}
\details{
The vector representation of a clustering is used for comparison
between clusterings.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

