% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{returnMetricsClust}
\alias{returnMetricsClust}
\title{Evaluates performance of inference of groups}
\usage{
returnMetricsClust(metricsIterTiMEx)
}
\arguments{
\item{metricsIterTiMEx}{structure with metrics of inference of the mutually
exclusive groups after running TiMEx iteratively, as returned by the 
function \code{clustMetrIterTiMEx}}
}
\value{
list consisting of 
\itemize{
\item{\code{percent}} {vector with length the number of real mutually 
exclusive pathways. each element represents the maximal percentage of the 
current group which was reconstructed by another group (real number between
0 and 1).} 
\item{\code{alone}} {vector with length the number of real mutually 
exclusive pathways. each element is either 0, if the initial group has not
been recovered by multiple groups, NA if it was and 2 if it hasn't been 
identified at all.}
}
}
\description{
\code{returnMetricsClust} returns metrics evaluating the 
performance in group recovery of running TiMEx iteratively
}
\details{

}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

