% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsCBN.R
\name{runCBN}
\alias{runCBN}
\title{Runs H-CBN}
\usage{
runCBN(path, name, optionsSA, noThreads)
}
\arguments{
\item{path}{path where the folder containing the initial results is created, 
as well as where the final results are written. In the H-CBN help file, 
this is the path where the \code{filestem} directory is created.}

\item{name}{name of the directory where results are initially written.}

\item{optionsSA}{string representing the command-line options of the
simmulated annealing optimization routine.}

\item{noThreads}{number of threads on which to run H-CBN.}
}
\value{
list consisting of 
\itemize{
\item{\code{optStruct}} {a square matrix of dimension input number of 
pathways plus 1, representing the inferred optimal structure. The last row
and column correspond to a dummy pathway and are identically zero.}
\item{\code{eps}} {H-CBN estimate for the gene-wise error rate (positive real 
number)}
\item{\code{alpha}} {H-CBN estimate for alpha (positive real number).}
\item{\code{loglik}} {H-CBN estimate for the log likelihood.}
\item{\code{lamobs}} {H-CBN estimate for the waiting time rate of the 
observation time (fixed to 1).}
\item{\code{lams}} {vector, H-CBN estimate for the waiting time rates of the 
events (pathways).}
}
}
\description{
\code{runCBN} runs the H-CBN program from the the command line, 
for already existing .pat (binary alteration matrix) and .poset (initial
solution for structure) files.
}
\details{
The folder were results of the H-CBN run are initially saved is 
\code{completeFilestem=path/name} (which, if not existent, is created). 
After running this function, five files are written: the optimized 
structure (poset), written to  \code{path/name/00000.poset} and also to 
\code{path/name.poset}, replacing the initial poset and becoming the initial
solution for the next iteration of the simmulated annealing optimization; 
the optimized lambda parameters, written to \code{path/name.lambda} and also
to \code{path/name/paramsEps.lambda}, where it also includes the estimate of
the error-rate epsilon; and the log of the simmulated annealing 
repetitions, which are written to \code{path/name.log}. The returned optimal
structure contains an additional dummy row and column identically set to 
0, in preparation for the format required by the local group optimization.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

