% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsSims.R
\name{sampleMEPathways}
\alias{sampleMEPathways}
\title{Draws samples from a mutually exclusive pathway}
\usage{
sampleMEPathways(m, TGenesTrunc, Tobs, muTrunc, eps)
}
\arguments{
\item{m}{number of samples to simulate, as a positive integer}

\item{TGenesTrunc}{matrix with waiting times of the genes; columns represent
the genes, rows represent the samples}

\item{Tobs}{vector of observation times, with length the number of samples}

\item{muTrunc}{intensity of mutual exclusivity of the pathway to be 
simulated, real number between 0 and 1}

\item{eps}{error rate at the level of genes, real number between 0 and 1}
}
\value{
a binary matrix representing the simulated mutually exclusive 
group, with columns as genes and rows as samples
}
\description{
\code{sampleMEPathways} simulates a group of mutually exclusive
genes, for given waiting times of genes, observation time, intensity of
mutual exclusivity and noise level.
}
\details{
In the case of perfect mutual exclusivity (\code{mu=1}), for each
sample, only the gene with the waiting time can be mutated, and only if its
waiting time is lower than the observation time.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

