% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsPoset.R
\name{topologicalSort}
\alias{topologicalSort}
\title{Computes the topological sort of a poset}
\usage{
topologicalSort(poset)
}
\arguments{
\item{poset}{the poset to be sorted}
}
\value{
a vector of events (vertices) topologically sorted
}
\description{
\code{topologicalSort} outputs the events in a poset according 
to their topological order.
}
\details{
the topological sorting of a poset is useful in iteratively 
going over all vertices in a poset and making sure every vertex is only
visited once
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

