% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsPoset.R
\name{trans_reduction}
\alias{trans_reduction}
\title{Transforms a poset into its transitive reduction}
\usage{
trans_reduction(poset)
}
\arguments{
\item{poset}{the input binary matrix representing the partial order between 
pathways}
}
\value{
a binary matrix, representing the transitive reduction of the input
poset
}
\description{
\code{trans_reduction} transforms a poset, i.e. a binary 
traiangular matrix corresponding to a directed acyclic graph, into another
binary matrix representing its transitive reduction.
}
\details{
the transitive reduction of a directed acyclic graph is the 
directed acyclic graph with as few edges as possible that has the same 
reachability relation as the initial graph.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

