% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsCBN.R
\name{transformAndWritePoset}
\alias{transformAndWritePoset}
\title{Transforms poset}
\usage{
transformAndWritePoset(path, name, oldStruct, newPathMat)
}
\arguments{
\item{path}{path where the new compatible poset is written}

\item{name}{name of the poset file to be written}

\item{oldStruct}{the old poset to be transformed, encoding order relations
between pathways at the previous optimization step}

\item{newPathMat}{current binary alteration matrix at the level of pathways, 
with rows representing samples and columns pathways, where an 1 represents 
that a pathway is altered in a sample}
}
\value{
nothing
}
\description{
\code{transformAndWritePoset} transforms the existing optimal
poset from the previous optimization step to a new compatible poset, to be
used as starting solution by the next run of H-CBN.
}
\details{
The role of this function is to create a compatible starting 
poset for the new H-CBN optimization step, on the basis of the current 
optimal poset. This is done by simply removing any edge from between events 
which do not longer exist as pathways. The numbering of the evnets is kept 
as in the current poset, without checking for consistency between pathways 
between the old and current groupings. The new poset is written to the file 
path/name.poset.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

