% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsCBN.R
\name{writePosetForCBN}
\alias{writePosetForCBN}
\title{Writes a poset to a file}
\usage{
writePosetForCBN(poset, path, name)
}
\arguments{
\item{poset}{the binary input matrix representing the order constraints
between events}

\item{path}{path where the output poset should be written. Default to
"~/Dropbox/PathTiMEx/RPathTiMEx/txts".}

\item{name}{name of the file to be written. Default to "outCBN".}
}
\value{
nothing
}
\description{
\code{writePosetForCBN} writes the input poset to a file, in 
a form accepted as input by H-CBN.
}
\details{
The poset encodes order constraints between events and it is either
used as starting solution for H-CBN, or the final result of the H-CBN 
structure otpimization. The poset is written to file in a form compatible to 
H-CBN, namely on the first line the number of events, on each of the 
following lines a directed edge representing an order constraint, and on 
the last line the number 0. If \code{path} and \code{name} are provided, 
the poset is written to \code{path/name.poset}.
}
\author{
Simona Constantinescu, \email{simona.constantinescu@bsse.ethz.ch}
}

